+++ tests/integration/chunkstore/test_chunkstore.py
@@ -1187,3 +1187,21 @@ def test_quarterly_data(chunkstore_lib):
         count += 1
 
     assert(count == 4)
+
+
+def test_list_symbols(chunkstore_lib):
+    df = DataFrame(data={'data': np.random.randint(0, 100, size=366)},
+                   index=pd.date_range('2016-01-01', '2016-12-31'))
+    df.index.name = 'date'
+
+    chunkstore_lib.write('rabbit', df)
+    chunkstore_lib.write('dragon', df)
+    chunkstore_lib.write('snake', df)
+    chunkstore_lib.write('wolf', df)
+    chunkstore_lib.write('bear', df)
+
+    assert('dragon' in chunkstore_lib.list_symbols())
+    assert(set(['rabbit', 'dragon', 'bear']) == set(chunkstore_lib.list_symbols(partial_match='r')))
+
+    assert(chunkstore_lib.has_symbol('dragon'))
+    assert(chunkstore_lib.has_symbol('marmot') is False)